+++ packages/components/psammead-media-player/src/Guidance/index.jsx
@@ -1,73 +1,51 @@
 import React from 'react';
-import { string, node } from 'prop-types';
+import { string } from 'prop-types';
 import styled from 'styled-components';
 import { C_WHITE } from '@bbc/psammead-styles/colours';
-import { mediaIcons } from '@bbc/psammead-assets/svgs';
-import {
-  GEL_SPACING_DBL,
-  GEL_SPACING_TRPL,
-  GEL_SPACING_HLF,
-} from '@bbc/gel-foundations/spacings';
+import { GEL_SPACING_DBL, GEL_SPACING } from '@bbc/gel-foundations/spacings';
 import { GEL_LONG_PRIMER } from '@bbc/gel-foundations/typography';
 import { getSansRegular } from '@bbc/psammead-styles/font-styles';
+import { GEL_GROUP_2_SCREEN_WIDTH_MIN } from '@bbc/gel-foundations/breakpoints';
 
 const GUIDANCE_BACKGROUND = 'rgba(34, 34, 34, 0.75)';
 
-const StyledGuidance = styled.div`
+const GuidanceWrapper = styled.div`
   ${({ service }) => getSansRegular(service)}
   ${GEL_LONG_PRIMER};
-`;
 
-const GuidanceWrapper = styled.div`
   width: 100%;
   height: 100%;
   position: absolute;
   background-color: ${GUIDANCE_BACKGROUND};
+  border: 0.0625rem solid transparent;
   color: ${C_WHITE};
-`;
-
-const Content = styled.div`
-  display: flex;
-  align-items: center;
-  padding: ${GEL_SPACING_DBL};
-`;
-
-const GuidanceMessage = styled.strong`
-  padding-left: ${GEL_SPACING_HLF};
-  font-weight: normal;
-`;
-
-const IconWrapper = styled.div`
-  > svg {
-    color: ${C_WHITE};
-    fill: currentColor;
-    height: ${GEL_SPACING_TRPL};
-    width: ${GEL_SPACING_TRPL};
+  @media screen and (-ms-high-contrast: active) {
+    background-color: transparent;
   }
 `;
 
-const Guidance = ({ guidanceMessage, service, children }) => {
-  return (
-    <StyledGuidance service={service}>
-      <GuidanceWrapper>
-        <Content>
-          <IconWrapper aria-hidden="true">{mediaIcons.guidance}</IconWrapper>
-          <GuidanceMessage>{guidanceMessage}</GuidanceMessage>
-        </Content>
-      </GuidanceWrapper>
-      {children}
-    </StyledGuidance>
-  );
-};
+const GuidanceMessage = styled.strong`
+  display: block;
+  font-weight: normal;
+  padding: ${GEL_SPACING};
+  border-bottom: 0.0625rem solid transparent;
+  @media screen and (-ms-high-contrast: active) {
+    background-color: window;
+  }
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
+    padding: ${GEL_SPACING_DBL};
+  }
+`;
+
+const Guidance = ({ guidanceMessage, service }) => (
+  <GuidanceWrapper service={service}>
+    <GuidanceMessage aria-hidden="true">{guidanceMessage}</GuidanceMessage>
+  </GuidanceWrapper>
+);
 
 Guidance.propTypes = {
   guidanceMessage: string.isRequired,
   service: string.isRequired,
-  children: node,
-};
-
-Guidance.defaultProps = {
-  children: null,
 };
 
 export default Guidance;